Fixed tests closure test
