+++ colibri/tests/test_utils.py
@@ -8,7 +8,6 @@ import shutil
 from numpy.testing import assert_allclose
 import pytest
 from unittest.mock import patch, mock_open, MagicMock
-from unittest import mock
 
 import jax
 import jax.numpy as jnp
@@ -23,7 +22,6 @@ from colibri.tests.conftest import (
 )
 from colibri.utils import (
     t0_pdf_grid,
-    closure_test_pdf_grid,
     resample_from_ns_posterior,
     cast_to_numpy,
     get_fit_path,
@@ -85,40 +83,6 @@ def test_t0_pdf_grid():
     N_fl = len(convolution.FK_FLAVOURS)  # number of flavours
 
     assert t0_grid.shape == (N_rep, N_fl, len(FIT_XGRID))
-
-
-def test_closure_test_pdf_grid():
-    """
-    Test the closure_test_pdf_grid function.
-
-    Verifies:
-    - Type of "closure_test_pdf" is validphys.core.PDF
-    - Output type is a jnp.array.
-    - The output shape is (N_rep, N_fl, N_x)
-    """
-    # mock a valid PDF set
-    inp = {"closure_test_pdf": "NNPDF40_nlo_as_01180"}
-    cltest_pdf_set = cAPI.closure_test_pdf(**inp)
-
-    # Check 1: closure_test_pdf is an instance of validphys.core.PDF
-    assert isinstance(cltest_pdf_set, validphys.core.PDF)
-
-    # define a test array
-    FIT_XGRID = [0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9]
-
-    # call the function
-    grid = closure_test_pdf_grid(cltest_pdf_set, FIT_XGRID, Q0=1.65)
-
-    # Check 2: type of the output is a jnp.array
-    assert isinstance(grid, jnp.ndarray)
-
-    # Check 3: shape of the output
-    N_rep = cltest_pdf_set.get_members()  #   number of replicas
-    N_fl = len(convolution.FK_FLAVOURS)  # number of flavours
-
-    print(f"FK_FLAVOURS: {convolution.FK_FLAVOURS}")
-
-    assert grid.shape == (N_rep, N_fl, len(FIT_XGRID))
 
 
 def test_resample_from_ns_posterior():
